Move legacy permission setting migration to legacy.py
