+++ src/yunohost/utils/legacy.py
@@ -4,8 +4,8 @@ from moulinette.utils.log import getActionLogger
 from moulinette.utils.filesystem import read_yaml
 
 from yunohost.user import user_list, user_group_create, user_group_update
-from yunohost.app import app_setting, _installed_apps
-from yunohost.permission import permission_create, user_permission_update, permission_sync_to_user
+from yunohost.app import app_setting, _installed_apps, _get_app_settings, _set_app_settings
+from yunohost.permission import permission_create, user_permission_list, user_permission_update, permission_sync_to_user
 
 logger = getActionLogger('yunohost.legacy')
 
@@ -145,3 +145,62 @@ LEGACY_PERMISSION_LABEL = {
 
 def legacy_permission_label(app, permission_type):
     return LEGACY_PERMISSION_LABEL.get((app, permission_type), "Legacy %s urls" % permission_type)
+
+
+def migrate_legacy_permission_settings(app=None):
+
+    logger.info(m18n.n("migrating_legacy_permission_settings"))
+    apps = _installed_apps()
+
+    if app:
+        if app not in apps:
+            logger.error("Can't migrate permission for app %s because it ain't installed..." % app)
+            apps = []
+        else:
+            apps = [app]
+
+    for app in apps:
+
+        settings = _get_app_settings(app) or {}
+
+        def _setting(name):
+            s = settings.get(name)
+            return s.split(',') if s else []
+
+        skipped_urls = [uri for uri in _setting('skipped_uris') if uri != '/']
+        skipped_urls += ['re:' + regex for regex in _setting('skipped_regex')]
+        unprotected_urls = [uri for uri in _setting('unprotected_uris') if uri != '/']
+        unprotected_urls += ['re:' + regex for regex in _setting('unprotected_regex')]
+        protected_urls = [uri for uri in _setting('protected_uris') if uri != '/']
+        protected_urls += ['re:' + regex for regex in _setting('protected_regex')]
+
+        if skipped_urls != []:
+            permission_create(app + ".legacy_skipped_uris", additional_urls=skipped_urls,
+                              auth_header=False, label=legacy_permission_label(app, "skipped"),
+                              show_tile=False, allowed='visitors', protected=True, sync_perm=False)
+        if unprotected_urls != []:
+            permission_create(app + ".legacy_unprotected_uris", additional_urls=unprotected_urls,
+                              auth_header=True, label=legacy_permission_label(app, "unprotected"),
+                              show_tile=False, allowed='visitors', protected=True, sync_perm=False)
+        if protected_urls != []:
+            permission_create(app + ".legacy_protected_uris", additional_urls=protected_urls,
+                              auth_header=True, label=legacy_permission_label(app, "protected"),
+                              show_tile=False, allowed=user_permission_list()['permissions'][app + ".main"]['allowed'],
+                              protected=True, sync_perm=False)
+
+        legacy_permission_settings = [
+            "skipped_uris",
+            "unprotected_uris",
+            "protected_uris",
+            "skipped_regex",
+            "unprotected_regex",
+            "protected_regex"
+        ]
+        for key in legacy_permission_settings:
+            if key in settings:
+                del settings[key]
+
+        _set_app_settings(app, settings)
+
+        permission_sync_to_user()
+
